Move name-archive-screen to `glam` folder
PER-9753: Glam Onboarding Header
