+++ src/app/onboarding/onboarding.module.ts
@@ -23,7 +23,7 @@ import { CreateNewArchiveComponent } from './components/create-new-archive/creat
 import { ArchiveTypeSelectComponent } from './components/archive-type-select/archive-type-select.component';
 import { ArchiveCreationStartScreenComponent } from './components/glam/archive-creation-start-screen/archive-creation-start-screen.component';
 import { SelectArchiveTypeScreenComponent } from './components/select-archive-type-screen/select-archive-type-screen.component';
-import { NameArchiveScreenComponent } from './components/name-archive-screen/name-archive-screen.component';
+import { NameArchiveScreenComponent } from './components/glam/name-archive-screen/name-archive-screen.component';
 import { GlamArchiveTypeSelectComponent } from './components/glam/archive-type-select/archive-type-select.component';
 import { ArchiveTypeIconComponent } from './components/glam/archive-type-icon/archive-type-icon.component';
 import { CreateArchiveForMeScreenComponent } from './components/glam/create-archive-for-me-screen/create-archive-for-me-screen.component';
